+++ css/style.css
@@ -806,6 +806,19 @@ main {
   background-color: #ffffff;
 }
 
+/*  * Сообщение сверху при ошибке отправки */
+
+.show-error {
+  z-index: 100;
+  position: absolute;
+  left: 0;
+  top: 0;
+  right: 0;
+  padding: 10px 3px;
+  font-size: 30px;
+  text-align: center;
+  background-color: red;
+}
 
 /*
   * Подпись к изображению
