+++ js/send.js
@@ -1,9 +1,7 @@
-import {showAlert} from './util.js';
+import {showAlert, renderPhotos} from './util.js';
 
-const pictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
-const pictureContainer = document.querySelector('.pictures');
-const pictureList = document.createDocumentFragment();
 const buttonSubmitForm = document.querySelector('.img-upload__form');
+const messageError = 'Не удалось отправить форму. Попробуйте ещё раз';
 
 const getRequest = () => {
   fetch('https://27.javascript.pages.academy/kekstagram-simple/data')
@@ -13,16 +11,11 @@ const getRequest = () => {
       }
       throw new Error(`${response.status} ${response.statusText}`);
     })
+
     .then((similarPictures) => {
-      similarPictures.forEach((image) => {
-        const pictureElement = pictureTemplate.cloneNode(true);
-        pictureElement.querySelector('.picture__img').src = image.url;
-        pictureElement.querySelector('.picture__likes').textContent = image.likes;
-        pictureElement.querySelector('.picture__comments').textContent = image.comments;
-        pictureList.appendChild(pictureElement);
-      });
-      pictureContainer.appendChild(pictureList);
+      renderPhotos(similarPictures);
     });
+
 };
 
 const setRequest = (onSuccess) => {
@@ -31,7 +24,7 @@ const setRequest = (onSuccess) => {
 
     const formData = new FormData(evt.target);
     fetch(
-      'https://27.javascript.pages.academy/kekstagram-simpl',
+      'https://27.javascript.pages.academy/kekstagram-simple',
       {
         method: 'POST',
         body: formData,
@@ -41,11 +34,11 @@ const setRequest = (onSuccess) => {
         if (response.ok) {
           onSuccess();
         } else {
-          showAlert('Не удалось отправить форму. Попробуйте ещё раз');
+          showAlert(messageError);
         }
       })
       .catch(() => {
-        showAlert('Не удалось отправить форму. Попробуйте ещё раз');
+        showAlert(messageError);
       });
   });
 };
